+++ components/sendFrameMetadata.tsx
@@ -1,0 +1,14 @@
+"use client"
+import { getFrameMetadata } from '@coinbase/onchainkit';
+
+const frameMetadata = getFrameMetadata({
+  buttons: [
+    {
+      label: "Begin"
+    }
+  ],
+  image: `${process.env.NEXT_PUBLIC_GATEWAY_URL}/ipfs/QmY24GiPcr2WSGhsRzWeREpVtsAkkvCY7e4ChYGpq2diVV/1.png`,
+  post_url: `${process.env.NEXT_PUBLIC_BASE_URL}/api/frame?id=1`,
+});
+
+export default frameMetadata;
