+++ package.json
@@ -10,9 +10,13 @@
   },
   "dependencies": {
     "@coinbase/onchainkit": "^0.14.0",
+    "@tanstack/react-query": "^5.35.1",
+    "graphql": "^16.8.1",
+    "graphql-request": "^6.1.0",
     "next": "^14.2.3",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react": "18",
+    "react-dom": "18",
+    "viem": "2.x"
   },
   "devDependencies": {
     "@types/node": "^20",
